tests/wpt/tests/cookiestore/cookieStore_delete.sub.https.html

<!doctype html>
<meta charset='utf-8'>
<title>Async Cookies: cookieStore basic API across origins</title>
<link rel='help' href='https://cookiestore.spec.whatwg.org/'>
<link rel='author' href='jarrydg@chromium.org' title='Jarryd Goodman'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='resources/helpers.js'></script>
<style>iframe { display: none; }</style>

<script>
'use strict';

const kPath = '/cookiestore/resources/helper_iframe.sub.html';
const kCorsBase = `https://{{domains[www1]}}:{{ports[https][0]}}`;
const kCorsUrl = `${kCorsBase}${kPath}`;

promise_test(async t => {
  const iframe = await createIframe(kCorsUrl, t);
  assert_true(iframe != null);

  iframe.contentWindow.postMessage({
    opname: 'set-cookie-without-domain-attr',
    name: 'cookie-name',
    value: 'cookie-value',
  }, kCorsBase);
  t.add_cleanup(async () => {
    await cookieStore.delete({ name: 'cookie-name', domain: '{{host}}' });
  });
  await waitForMessage();

  iframe.contentWindow.postMessage({
    opname: 'get-cookie',
    name: 'cookie-name',
    options: {}
  }, '*');

  {
    const message = await waitForMessage();
    const { frameCookie } = message;
    assert_not_equals(frameCookie, null);
    assert_equals(frameCookie.name, 'cookie-name');
    assert_equals(frameCookie.value, 'cookie-value');
  }

  iframe.contentWindow.postMessage({
    opname: 'delete-cookie-without-domain-attr',
    name: 'cookie-name',
  }, kCorsBase);

  await waitForMessage();

  iframe.contentWindow.postMessage({
    opname: 'get-cookie',
    name: 'cookie-name',
    options: {}
  }, '*');

  {
    const message = await waitForMessage();
    const { frameCookie } = message;
    assert_equals(frameCookie, null);
  }
});
</script>
